requirements.txt
# Data & API
requests==2.31.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Scheduling
schedule==1.2.0

# Email & Newsletter
jinja2==3.1.2

# CLI
click==8.1.7
rich==13.7.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Jira Integration
jira==3.5.2
